1.21.5 port

+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/tablet/TabletEntry.java
@@ -43,10 +43,10 @@ public record TabletEntry(String id, String description, ResourceLocation icon, 
         ).apply(instance, Image::new));
     }
 
-    public record Item(ItemStack stack, float size, Optional<Boolean> onlyIcon) {
+    public record Item(ItemStack stack, int size, Optional<Boolean> onlyIcon) {
         public static final Codec<Item> CODEC = RecordCodecBuilder.create(instance -> instance.group(
                 ItemStack.CODEC.fieldOf("stack").forGetter(Item::stack),
-                Codec.FLOAT.fieldOf("size").forGetter(Item::size),
+                Codec.INT.fieldOf("size").forGetter(Item::size),
                 Codec.BOOL.optionalFieldOf("onlyIcon").forGetter(Item::onlyIcon)
         ).apply(instance, Item::new));
     }
